+++ src/components/login-register-style.css
